Added odeUtility to make array operations easier. Added tests for it.
