+++ package.json
@@ -22,11 +22,11 @@
         "react-transition-group": "^4.4.2",
         "typewriter-effect": "^2.18.2",
         "web-vitals": "^3.1.1"
-    },
+    }, 
     "scripts": {
         "start": "react-scripts start",
         "build": "react-scripts build",
-        "test": "react-scripts test",
+        "test": "react-scripts test --setupFilesAfterEnv='./setupTests.js'",
         "eject": "react-scripts eject",
         "predeploy": "yarn build && cp build/index.html build/404.html",
         "deploy": "gh-pages -d build"
